+++ pkg/schema/src/auth/queries.ts
