+++ pkg/schema/src/challenge/responses.ts
