+++ pkg/schema/src/common/queries.ts
@@ -1,0 +1,7 @@
+import { Type, Static } from '@sinclair/typebox';
+
+export const PaginatedRequest = Type.Object({
+  size: Type.Number({ default: 25, maximum: 50, minimum: 0 }),
+  after: Type.Number({ default: 0 }),
+});
+export type PaginatedRequestType = Static<typeof PaginatedRequest>;
