+++ pkg/schema/src/common/responses.ts
@@ -5,12 +5,6 @@ export const ErrorResponse = Type.Object({
 });
 export type ErrorResponseType = Static<typeof ErrorResponse>;
 
-export const PaginatedRequest = Type.Object({
-  size: Type.Number({ default: 25, maximum: 50, minimum: 0 }),
-  after: Type.Number({ default: 0 }),
-});
-export type PaginatedRequestType = Static<typeof PaginatedRequest>;
-
 export const PaginatedResponse = <T extends TSchema>(type: T) => Type.Object({
   page: type, // Our Actual Data
   cursor: Type.Object({
