--- pkg/schema/src/index.ts
@@ -1,4 +1,0 @@
-export * as datatypes from './datatypes';
-export * as queries from './queries';
-export * as requests from './requests';
-export * as responses from './responses';
